requirements.txt
# Projects dependencies

# Image processing
pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0

# User interface
streamlit==1.12.0
altair<5

# Model dependencies
huggingface-hub==0.20.3
accelerate==0.27.2

# Others
typing
